pyproject.toml
[project]
name = "wanderline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy",
    "opencv-python",
    "scikit-image",
    "pillow",
    "pytest>=8.4.1",
]

[tool.poetry.dev-dependencies]
pytest = "^7.0"
